+++ packages/connection/tests/php/test_Manager.php
